+++ src/features/speedrun/season2/constants.ts
@@ -1,7 +1,6 @@
 import { range } from "isaacscript-common";
+import g from "../../../globals";
 import { CollectibleTypeCustom } from "../../../types/CollectibleTypeCustom";
-
-export const SEASON_2_DEBUG = false;
 
 export const SEASON_2_NUM_BANS = 3;
 
@@ -187,6 +186,6 @@ export const SEASON_2_FORGOTTEN_EXCEPTIONS: readonly int[] =
   });
 
 /** How long the randomly-selected character & build combination is "locked-in". */
-const SEASON_2_LOCK_MINUTES = 1.25;
+const SEASON_2_LOCK_MINUTES = g.debug ? 0.01 : 1.25;
 export const SEASON_2_LOCK_SECONDS = SEASON_2_LOCK_MINUTES * 60;
 export const SEASON_2_LOCK_MILLISECONDS = SEASON_2_LOCK_SECONDS * 1000;